+++ CHANGELOG.md
@@ -12,6 +12,7 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 
 ### Added
 
+* [Patterns] First pattern for 'page not found' pages, with associated Testbed changes
 * [Styles] Source EPS files for BAS roundal and logo
 * [Styles] BAS Favicon based on the inverse BAS roundel
 * [Scripts] Sortable list/table component using HTML5 Sortable